test(Effect): consolidate repeatOrElse coverage
